+++ R/thicken_cust.R
@@ -35,6 +35,7 @@ thicken_cust <- function(x,
                          drop = FALSE) {
 
   is_df(x)
+  has_rows(x)
 
   original_data_frame <- x
   x <- as.data.frame(x)