+++ tests/testthat/test_thicken.R
@@ -146,7 +146,25 @@ Returned dataframe contains original observations, with NA values for d and d_we
 
   expect_equal(sw(thicken(coffee_two_nas, "week"))$d_week,
                c(as.Date("2016-07-03"), NA, NA, as.Date("2016-07-10")))
+})
 
+test_that("add_na_to_thicken unit tests", {
+  thickened_date     <- ymd(c(20180101, 20180101))
+  thickened_datetime <- ymd_h(c("20180101 01", "20180101 04"))
+  thickened_datetime_CET <- as.POSIXct(as.character(thickened_datetime), tz = "CET")
+
+
+  expect_equal(add_na_to_thicken(thickened_date, 2),
+               ymd(c(20180101, NA, 20180101)))
+  expect_equal(add_na_to_thicken(thickened_datetime, 2),
+               ymd_h(c("20180101 01", NA, "20180101 04")))
+  expect_equal(add_na_to_thicken(thickened_date, c(2, 4)),
+               ymd(c(20180101, NA, 20180101, NA)))
+  expect_equal(add_na_to_thicken(thickened_datetime, c(2, 4)),
+               ymd_h(c("20180101 01", NA, "20180101 04", NA)))
+
+  expect_equal(add_na_to_thicken(thickened_datetime_CET, c(2, 4)),
+               ymd_h(c("20180101 01", NA, "20180101 04", NA), tz = "CET"))
 })
 
 context("thicken drop argument")